Divide classes into suitable java packages
