+++ src/main/java/Duke.java
@@ -1,3 +1,10 @@
+import duke.command.Command;
+import duke.exception.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 import java.nio.file.Paths;
 import java.time.format.DateTimeParseException;
