+++ src/main/java/duke/command/AddCommand.java
@@ -1,4 +1,14 @@
+package duke.command;
+
 import java.time.LocalDate;
+
+import duke.storage.Storage;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.Todo;
+import duke.ui.Ui;
 
 public class AddCommand extends Command {
