+++ src/main/java/duke/exception/EmptyTaskDescriptionException.java
@@ -1,0 +1,8 @@
+package duke.exception;
+
+public class EmptyTaskDescriptionException extends DukeException {
+
+    public EmptyTaskDescriptionException(String taskType) {
+        super("The description of a " + taskType + " cannot be empty!");
+    }
+}
