+++ src/main/java/duke/exception/MissingTaskIndexException.java
@@ -1,0 +1,9 @@
+package duke.exception;
+
+public class MissingTaskIndexException extends DukeException {
+
+    public MissingTaskIndexException() {
+        super("Task index is not specified!\n"
+                + "Please follow this format: [done/delete] [index]");
+    }
+}
