+++ src/main/java/duke/exception/TaskAlreadyDoneException.java
@@ -1,0 +1,8 @@
+package duke.exception;
+
+public class TaskAlreadyDoneException extends DukeException {
+
+    public TaskAlreadyDoneException() {
+        super("This task has already been done!");
+    }
+}
