+++ src/main/java/duke/exception/TaskIndexOutOfBoundsException.java
@@ -1,0 +1,8 @@
+package duke.exception;
+
+public class TaskIndexOutOfBoundsException extends DukeException {
+
+    public TaskIndexOutOfBoundsException() {
+        super("Invalid task index specified!");
+    }
+}
