+++ src/main/java/duke/exception/UnrecognisedCommandException.java
@@ -1,0 +1,8 @@
+package duke.exception;
+
+public class UnrecognisedCommandException extends DukeException {
+
+    public UnrecognisedCommandException() {
+        super("This command is not recognised! Please enter a valid command!");
+    }
+}
