+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,16 @@
+package duke.parser;
+
+import duke.command.AddCommand;
+import duke.command.Command;
+import duke.command.DeleteCommand;
+import duke.command.DoneCommand;
+import duke.command.ExitCommand;
+import duke.command.ListCommand;
+import duke.exception.EmptyTaskDescriptionException;
+import duke.exception.MissingTaskIndexException;
+import duke.exception.TimeNotSpecifiedException;
+import duke.exception.UnrecognisedCommandException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
