+++ src/main/java/duke/task/Todo.java
@@ -1,5 +1,7 @@
+package duke.task;
+
 /**
- * This Todo class implements the characteristics of a task
+ * This duke.task.Todo class implements the characteristics of a task
  * without any date or time attached to it.
  *
  * @author Yeo Jun Wei
@@ -7,7 +9,7 @@
 public class Todo extends Task {
 
     /**
-     * Constructor for a Todo instance that takes in a description.
+     * Constructor for a duke.task.Todo instance that takes in a description.
      *
      * @param description The given task description.
      */
@@ -16,9 +18,9 @@ public class Todo extends Task {
     }
 
     /**
-     * Returns the string representation of a Todo instance.
+     * Returns the string representation of a duke.task.Todo instance.
      *
-     * @return A string representing a Todo instance.
+     * @return A string representing a duke.task.Todo instance.
      */
     @Override
     public String toString() {
